requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
pycryptodome==3.21.0

